Splitting toJapanese and toArabic code for lighter integrations
