+++ lib/calc.js
