+++ lib/factors.js
@@ -1,0 +1,21 @@
+'use strict';
+module.exports = Object.freeze([
+    {j: '万', a: '10000'},
+    {j: '億', a: '100000000'},
+    {j: '兆', a: '1000000000000'},
+    {j: '京', a: '10000000000000000'},
+    {j: '垓', a: '100000000000000000000'},
+    // {j: '𥝱', a: '1000000000000000000000000'},
+    {j: '秭', a: '1000000000000000000000000'},
+    {j: '穣', a: '10000000000000000000000000000'},
+    {j: '溝', a: '100000000000000000000000000000000'},
+    {j: '澗', a: '1000000000000000000000000000000000000'},
+    {j: '正', a: '10000000000000000000000000000000000000000'},
+    {j: '載', a: '100000000000000000000000000000000000000000000'},
+    {j: '極', a: '1000000000000000000000000000000000000000000000000'},
+    {j: '恒河沙', a: '10000000000000000000000000000000000000000000000000000'},
+    {j: '阿僧祇', a: '100000000000000000000000000000000000000000000000000000000'},
+    {j: '那由他', a: '1000000000000000000000000000000000000000000000000000000000000'},
+    {j: '不可思議', a: '10000000000000000000000000000000000000000000000000000000000000000'},
+    {j: '無量大数', a: '100000000000000000000000000000000000000000000000000000000000000000000'},
+]);
