+++ lib/reverse.js
@@ -9,4 +9,3 @@
 module.exports = function reverse(array) {
     return Array.from(array).reverse();
 };
-
